requirements.txt
python-telegram-bot==20.3
Pillow==10.3.0
pytesseract==0.3.10
fastapi==0.115.12
uvicorn==0.34.2